+++ app/configurator/table/configurator/links-section.tsx
@@ -1,9 +1,8 @@
-import { sanitizeUrl } from "@braintree/sanitize-url";
 import { t, Trans } from "@lingui/macro";
 import { SelectChangeEvent } from "@mui/material";
-import { KeyboardEvent, useEffect, useMemo, useState } from "react";
+import { useMemo } from "react";
 
-import { Input, Select } from "@/components/form";
+import { Select } from "@/components/form";
 import { TableConfig } from "@/config-types";
 import {
   ControlSection,
@@ -16,8 +15,10 @@ import {
   useConfiguratorState,
 } from "@/configurator/configurator-state";
 import { Dimension } from "@/domain/data";
-import { useLocale } from "@/locales/use-locale";
+import { useLocale, useOrderedLocales } from "@/locales/use-locale";
 import { useEvent } from "@/utils/use-event";
+
+import { LinkBaseUrlInput } from "./link-base-url-input";
 
 export const TableLinksSection = ({
   chartConfig,
@@ -27,6 +28,7 @@ export const TableLinksSection = ({
   dimensions: Dimension[];
 }) => {
   const locale = useLocale();
+  const orderedLocales = useOrderedLocales();
   const [_, dispatch] = useConfiguratorState(isConfiguring);
 
   const dimensionOptions = useMemo(() => {
@@ -92,10 +94,15 @@ export const TableLinksSection = ({
           field={null}
           path="links.enabled"
         />
-        <BaseUrlInput
-          value={chartConfig.links.baseUrl}
-          disabled={!chartConfig.links.enabled}
-        />
+        {orderedLocales.map((locale) => (
+          <div key={`${locale}-table-link`}>
+            <LinkBaseUrlInput
+              value={chartConfig.links.baseUrl[locale]}
+              disabled={!chartConfig.links.enabled}
+              locale={locale}
+            />
+          </div>
+        ))}
         <Select
           id="links.componentId"
           size="sm"
@@ -125,99 +132,4 @@ export const TableLinksSection = ({
       </ControlSectionContent>
     </ControlSection>
   );
-};
-
-const BaseUrlInput = ({
-  value,
-  disabled,
-}: {
-  value: string;
-  disabled: boolean;
-}) => {
-  const locale = useLocale();
-  const [_, dispatch] = useConfiguratorState(isConfiguring);
-
-  const [inputValue, setInputValue] = useState(value);
-  const [isValid, setIsValid] = useState(true);
-
-  useEffect(() => {
-    setInputValue(value);
-  }, [value]);
-
-  const updateBaseUrl = useEvent((newValue: string) => {
-    dispatch({
-      type: "CHART_FIELD_UPDATED",
-      value: {
-        locale,
-        field: null,
-        path: "links.baseUrl",
-        value: newValue,
-      },
-    });
-  });
-
-  const handleCommit = useEvent(() => {
-    if (inputValue === "") {
-      setIsValid(true);
-      updateBaseUrl("");
-
-      return;
-    }
-
-    const sanitizedUrl = sanitizeUrl(inputValue);
-
-    if (sanitizedUrl === "about:blank") {
-      setIsValid(false);
-      updateBaseUrl("");
-
-      return;
-    }
-
-    try {
-      const url = new URL(sanitizedUrl);
-      const normalizedUrl = normalizeUrl(url);
-
-      updateBaseUrl(normalizedUrl);
-      setIsValid(true);
-      setInputValue(normalizedUrl);
-    } catch {
-      setIsValid(false);
-    }
-  });
-
-  const handleKeyDown = useEvent((e: KeyboardEvent<HTMLInputElement>) => {
-    if (e.key === "Enter") {
-      handleCommit();
-    }
-  });
-
-  return (
-    <Input
-      type="url"
-      label={t({
-        id: "controls.tableSettings.baseUrl",
-        message: "Base URL",
-      })}
-      name="links.baseUrl"
-      placeholder="https://example.com/"
-      value={inputValue}
-      disabled={disabled}
-      error={!isValid}
-      errorMessage={t({
-        id: "controls.tableSettings.baseUrlInvalid",
-        message: "Please enter a valid URL",
-      })}
-      onChange={(e) => setInputValue(e.target.value)}
-      onBlur={handleCommit}
-      onKeyDown={handleKeyDown}
-    />
-  );
-};
-
-const normalizeUrl = (url: URL) => {
-  if (!url.pathname.endsWith("/")) {
-    url.pathname = url.pathname + "/";
-  }
-
-  return url.toString();
 };